+++ tripFrontend/src/components/Error/FieldErrorCustom/index.ts
@@ -1,2 +1,1 @@
-export * from './ErrorText';
 export * from './FieldErrorCustom';
