+++ tripFrontend/src/components/Form/Row/Row.tsx
@@ -1,0 +1,26 @@
+import { JSX } from 'react';
+
+type TRow = {
+	label: string;
+	required?: boolean;
+	children: React.ReactNode;
+	top?: boolean;
+};
+
+const Required = (): JSX.Element => (
+	<span className="text-red-600" aria-hidden>
+		*
+	</span>
+);
+
+export const Row = ({ label, required, children, top }: TRow): JSX.Element => (
+	<div
+		className={`grid grid-cols-[260px_1fr] ${top ? 'items-start' : 'items-center'} mb-4`}
+	>
+		<label className="font-medium text-gray-800">
+			{label}
+			{required && <Required />}
+		</label>
+		{children}
+	</div>
+);
