+++ tripFrontend/src/components/Form/Textarea/index.ts
@@ -1,0 +1,1 @@
+export * from './Textarea';
