+++ tripFrontend/src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './ComboBox';
 export * from './Error';
 export * from './Footer';
+export * from './Form';
 export * from './Header';
 export * from './SelectBox';
 export * from './TableView';
