+++ tripFrontend/src/features/product/components/ProductForm/ProductForm.tsx
@@ -3,16 +3,10 @@
 import { isCuid } from 'cuid';
 import { X, Save, Plus, Map, Trash2 } from 'lucide-react';
 import { JSX, useState, useEffect, useMemo } from 'react';
-import {
-	FieldErrors,
-	SubmitHandler,
-	useForm,
-	UseFormReturn,
-	UseFormSetValue,
-} from 'react-hook-form';
+import { SubmitHandler, UseFormReturn } from 'react-hook-form';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { ComboBox } from '@/components';
+import { Input, Row, Select, Textarea } from '@/components';
 import { DiscountBoard } from '@/features/discount';
 import { locationThunk } from '@/features/location';
 import {
@@ -26,53 +20,6 @@ import { TDiscountDetail, TProductSchedule } from '@/types';
 
 import { TRequestBodyCreateProduct } from '../../product.type';
 
-type FieldErr = FieldErrors<TRequestBodyCreateProduct>;
-
-type TError = {
-	id: keyof TRequestBodyCreateProduct;
-	errors: FieldErr;
-};
-
-type TRow = {
-	label: string;
-	required?: boolean;
-	children: React.ReactNode;
-	top?: boolean;
-};
-
-type TInputProps = {
-	id: keyof TRequestBodyCreateProduct;
-	label: string;
-	defaultValue?: string;
-	register: ReturnType<typeof useForm<TRequestBodyCreateProduct>>['register'];
-	errors: FieldErr;
-	required?: boolean;
-	disabled?: boolean;
-	type?: string;
-	validate?: (value: string | number | boolean) => string | boolean;
-};
-
-type TTextareaProps = {
-	id: keyof TRequestBodyCreateProduct;
-	label: string;
-	register: ReturnType<typeof useForm<TRequestBodyCreateProduct>>['register'];
-	required?: boolean;
-	errors: FieldErr;
-};
-
-type TSelect = {
-	id: keyof TRequestBodyCreateProduct;
-	label: string;
-	setValue: UseFormSetValue<TRequestBodyCreateProduct>;
-	errors: FieldErr;
-	options: { id: string; value: string; label: string; city?: string }[];
-	required?: boolean;
-	placeholder?: string;
-	disabled?: boolean;
-	defaultValue?: string;
-	name?: keyof TRequestBodyCreateProduct;
-};
-
 type TProductFormProps = {
 	form: UseFormReturn<TRequestBodyCreateProduct>;
 	schedules?: TProductSchedule[] | TRequestBodyCreateSchedule[];
@@ -84,127 +31,6 @@ type TProductFormProps = {
 	onRemove?: () => void;
 	onSubmit?: SubmitHandler<TRequestBodyCreateProduct>;
 	onCancel?: () => void;
-};
-
-const Required = (): JSX.Element => (
-	<span className="text-red-600" aria-hidden>
-		*
-	</span>
-);
-
-const ErrorText = ({ id, errors }: TError): JSX.Element => (
-	<p id={`${id}-error`} className="min-h-[1.25rem] text-sm text-red-600">
-		{errors[id]?.message as string}
-	</p>
-);
-
-const Row = ({ label, required, children, top }: TRow): JSX.Element => (
-	<div
-		className={`grid grid-cols-[260px_1fr] ${top ? 'items-start' : 'items-center'} mb-4`}
-	>
-		<label className="font-medium text-gray-800">
-			{label}
-			{required && <Required />}
-		</label>
-		{children}
-	</div>
-);
-
-const Input = ({
-	id,
-	label,
-	defaultValue,
-	register,
-	errors,
-	required,
-	disabled = false,
-	type = 'text',
-	validate,
-}: TInputProps): JSX.Element => (
-	<Row label={label} required={required}>
-		<div>
-			<input
-				id={id as string}
-				type={type}
-				aria-invalid={errors[id] ? 'true' : undefined}
-				aria-describedby={`${id}-error`}
-				defaultValue={defaultValue}
-				disabled={disabled}
-				className={`h-10 w-full border-b border-gray-300 focus:border-gray-500 focus:outline-none ${disabled ? 'bg-gray-100 border-none pl-2.5' : 'bg-white'}`}
-				{...register(id, {
-					...(required && { required: `${label} is required` }),
-					...(type === 'number' && { valueAsNumber: true }),
-					...(validate && { validate: validate }),
-					setValueAs: (value) => value.trim(),
-				})}
-			/>
-			<ErrorText id={id} errors={errors} />
-		</div>
-	</Row>
-);
-
-const Textarea = ({
-	id,
-	label,
-	required,
-	register,
-	errors,
-}: TTextareaProps): JSX.Element => (
-	<Row label={label} top>
-		<div>
-			<textarea
-				id={id as string}
-				className="h-[100px] w-full rounded-md bg-white-100 p-4 border border-gray-300 focus:outline-none"
-				{...register(id, {
-					...(required && { required: `${label} is required` }),
-					setValueAs: (value) => value.trim(),
-				})}
-			/>
-			<ErrorText id={id} errors={errors} />
-		</div>
-	</Row>
-);
-
-const Select = ({
-	id,
-	label,
-	setValue,
-	errors,
-	options,
-	required,
-	placeholder,
-	disabled = false,
-	defaultValue,
-	name,
-}: TSelect): JSX.Element => {
-	const [selectedOption, setSelectedOption] = useState<{
-		id: string;
-		value: string;
-		label: string;
-		city?: string;
-	} | null>(null);
-
-	useEffect(() => {
-		const option = options.find((o) => o.label === defaultValue);
-		if (option) {
-			setSelectedOption(option);
-		}
-		const fieldName = name || id;
-		setValue(fieldName, selectedOption?.value || '');
-	}, [options, defaultValue, selectedOption, setValue, name, id]);
-
-	return (
-		<Row label={label} required={required}>
-			<ComboBox
-				options={options}
-				selectedOption={selectedOption}
-				setSelectedOption={setSelectedOption}
-				placeholder={placeholder}
-				disabled={disabled}
-			/>
-			<ErrorText id={id} errors={errors} />
-		</Row>
-	);
 };
 
 export const ProductForm = ({
@@ -361,14 +187,14 @@ export const ProductForm = ({
 					className="px-6 pb-8 pt-6"
 					onSubmit={onSubmit && handleSubmit(onSubmit)}
 				>
-					<Input
+					<Input<TRequestBodyCreateProduct>
 						id="name"
 						label="Product Name"
 						required
 						register={register}
 						errors={errors}
 					/>
-					<Select
+					<Select<TRequestBodyCreateProduct>
 						id="locationId"
 						label="Location On System"
 						required
@@ -378,14 +204,14 @@ export const ProductForm = ({
 						options={options}
 						placeholder="Select location"
 					/>
-					<Textarea
+					<Textarea<TRequestBodyCreateProduct>
 						id="description"
 						label="Description"
 						register={register}
 						errors={errors}
 						required
 					/>
-					<Input
+					<Input<TRequestBodyCreateProduct>
 						id="cityName"
 						label="Destination"
 						defaultValue={watch('cityName')}
@@ -394,7 +220,7 @@ export const ProductForm = ({
 						// required
 						disabled
 					/>
-					<Input
+					<Input<TRequestBodyCreateProduct>
 						id="time"
 						label="Time (Hour)"
 						register={register}
@@ -403,7 +229,7 @@ export const ProductForm = ({
 						type="number"
 						validate={validateGreaterThanZero}
 					/>
-					<Input
+					<Input<TRequestBodyCreateProduct>
 						id="quantityAvailable"
 						label="Quantity (Person)"
 						register={register}
@@ -412,7 +238,7 @@ export const ProductForm = ({
 						type="number"
 						validate={validateGreaterThanZero}
 					/>
-					<Input
+					<Input<TRequestBodyCreateProduct>
 						id="age"
 						label="Age"
 						register={register}
@@ -479,7 +305,9 @@ export const ProductForm = ({
 										}
 									/>
 								) : (
-									title === 'Discounts' && <DiscountBoard data={discounts} />
+									title === 'Discounts' && (
+										<DiscountBoard data={discounts} />
+									)
 								)}
 							</div>
 						</section>
